Apps and Config enhancements
added the capability to rename and clone apps
added the ability to add env vars to the global app and to unset and toggle restart when setting env vars for all apps. if a key is both in the config dict AND in the unset list, then env var will first be added and then deleted. all of these functions will respect the restart instruction.