feat: make version pinning optional
The ansible role was pinning the versions of dokku-related apt packages.
In order to use the role with newer (or older) dokku versions, users of
the role would have to figure out by themselves which versions of these
packages were compatible - despite the fact that a simple `apt-get
install dokku` would have taken care of the dependencies correctly.

Here we remove version pinning by default, while still giving users the
option to pin the version of certain packages if they have a need to do
so.