fix package pinning
After PR 97, package pinning was not reflected in the packages actually
installed since the ansible `apt` module with `state: present` does not
ensure that packages comply with a pinned version - it only makes sure
the package is present.

In order to install the pinned version, the version needs to be passed
to the `apt` module explicitly.
This was the case prior to PR 97 and is reintroduced here.